refactor(tools): move BashTool back into cube-standard core
BashTool has zero extra dependencies (only uses Container and Tool from
cube-standard itself), so the optional-package overhead buys nothing.
It belongs in core alongside the Container it wraps.

- Moves bash.py back to src/cube/tools/bash.py
- Re-exports BashTool, BashToolConfig from cube.tools
- Removes cube-tools/cube-bash-tool/ package
- Removes entry from cube-tools/README.md
