--- cube-tools/cube-bash-tool/pyproject.toml
@@ -1,43 +1,0 @@
-[project]
-name = "cube-bash-tool"
-version = "0.1.0"
-description = "Bash tool implementation for cube-standard benchmarks"
-requires-python = ">=3.12"
-dependencies = [
-    "cube-standard",
-]
-
-[project.optional-dependencies]
-dev = [
-    "pytest>=8.0.0",
-    "pytest-asyncio>=0.24.0",
-]
-
-[tool.uv.sources]
-cube-standard = { path = "../..", editable = true }
-
-[build-system]
-requires = ["uv_build>=0.6.0,<0.7.0"]
-build-backend = "uv_build"
-
-[tool.uv.build-backend]
-module-name = "cube_bash_tool"
-
-[tool.ruff]
-fix = true
-line-length = 120
-indent-width = 4
-
-[tool.ruff.format]
-quote-style = "double"
-indent-style = "space"
-skip-magic-trailing-comma = false
-line-ending = "auto"
-
-[tool.ruff.lint]
-extend-select = ["I"]
-
-[tool.pytest.ini_options]
-testpaths = ["tests"]
-addopts = "-rs"
-asyncio_mode = "auto"
